requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.4
django-allauth==0.62.1
django-appconf==1.0.6
django-compressor==4.4
django-extensions==3.2.3
pillow==10.3.0
rcssmin==1.1.1
reportlab==4.2.0
rjsmin==1.2.1
sqlparse==0.5.0
tzdata==2024.1
